+++ internal/repository/component.go
@@ -110,11 +110,6 @@ func (r *HostComponentRepository) GetComponent(name string) (*components.Compone
 			return err
 		}
 
-		if newRes.Kind == components.ResourceTypeComponentScript {
-			scripts++
-			oldComp.Scripted = true
-		}
-
 		return nil
 	})
 	if err != nil {
@@ -202,7 +197,7 @@ func (r *HostComponentRepository) ListResources(c *components.Component) ([]comp
 			newRes := components.Resource{
 				Parent:   c.Name,
 				Path:     resName,
-				Kind:     c.FindResourceType(resName),
+				Kind:     components.FindResourceType(resName),
 				Template: components.IsTemplate(resName),
 			}
 			resources = append(resources, newRes)
@@ -341,7 +336,7 @@ func (r *HostComponentRepository) NewResource(parent *components.Component, path
 	if fileInfo.IsDir() {
 		res.Kind = components.ResourceTypeDirectory
 	} else {
-		res.Kind = parent.FindResourceType(path)
+		res.Kind = components.FindResourceType(path)
 	}
 	if res.IsQuadlet() {
 		res.Path, err = filepath.Rel(filepath.Join(r.QuadletPrefix, parent.Name), path)
@@ -362,11 +357,6 @@ func (r *HostComponentRepository) NewResource(parent *components.Component, path
 		res.Path, err = filepath.Rel(filepath.Join(r.DataPrefix, parent.Name), path)
 		if err != nil {
 			return res, err
-		}
-	}
-	if res.Kind == components.ResourceTypeComponentScript {
-		if fileInfo.Mode().Perm()&0o111 != 0o111 {
-			return res, fmt.Errorf("component script %v is not executable", path)
 		}
 	}
 	return res, nil