+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 Best effort list of major changes and bugfixes
 
 ## Upcoming
+- feat: Component scripts are no longer a separate resource type. They are instead normal script resources that, when specified in the component manifest, are run in a transient systemd unit on installation/removal.
 
 ## 0.5.1
 - bugfix: fixed some typos around volume importing